iPochta-back/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
decorator==5.1.1
defusedxml==0.7.1
dependency-injector==4.41.0
dnspython==2.7.0
docopt==0.6.2
email_validator==2.1.1
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.115.3
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
fastjsonschema==2.19.1
frozenlist==1.5.0
gevent==24.10.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.4.2
magic-filter==1.0.12
makefun==1.15.6
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
packaging==24.0
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
propcache==0.2.0
ptyprocess==0.7.0
pure-eval==0.2.2
pwdlib==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
pyzmq==25.1.2
redis==5.2.0
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.2
urllib3==2.2.1
uvicorn==0.27.0.post1
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
websockets==13.1
yarg==0.1.9
yarl==1.16.0
zope.event==5.0
zope.interface==7.1.1
